Move content to 'component' folder and add some messages to 'app'
