+++ generators/component/templates/component-index.js
