+++ generators/component/templates/component-js.js
