+++ generators/component/templates/component-template.slim
